Show and Hide Working
Reorganized the html with new div to help with show and hide feature. Updated the server side get index route so it only return the user's index (instead of the whole index). Rework some of the promises in events to implement show and hide (similar work done in the app.js). Reword some button and response messages.